+++ libraries/base/Prelude.hs
@@ -66,11 +66,32 @@ module Prelude (
     subtract, even, odd, gcd, lcm, (^), (^^),
     fromIntegral, realToFrac,
 
+    -- ** Monoids
+    Monoid(mempty, mappend, mconcat),
+
     -- ** Monads and functors
     Functor(fmap),
     Applicative(pure, (<*>), (*>), (<*)),
     Monad((>>=), (>>), return, fail),
-    mapM, mapM_, sequence, sequence_, (=<<),
+    mapM_, sequence_, (=<<),
+
+    -- ** Folds and traversals
+    Foldable(elem,      -- :: (Foldable t, Eq a) => a -> t a -> Bool
+             -- fold,   -- :: Monoid m => t m -> m
+             foldMap,   -- :: Monoid m => (a -> m) -> t a -> m
+             foldr,     -- :: (a -> b -> b) -> b -> t a -> b
+             -- foldr', -- :: (a -> b -> b) -> b -> t a -> b
+             foldl,     -- :: (b -> a -> b) -> b -> t a -> b
+             -- foldl', -- :: (b -> a -> b) -> b -> t a -> b
+             foldr1,    -- :: (a -> a -> a) -> t a -> a
+             foldl1,    -- :: (a -> a -> a) -> t a -> a
+             maximum,   -- :: (Foldable t, Ord a) => t a -> a
+             minimum,   -- :: (Foldable t, Ord a) => t a -> a
+             product,   -- :: (Foldable t, Num a) => t a -> a
+             sum),      -- :: Num a => t a -> a
+             -- toList) -- :: Foldable t => t a -> [a]
+
+    Traversable(traverse, sequenceA, mapM, sequence),
 
     -- ** Miscellaneous functions
     id, const, (.), flip, ($), until,
@@ -81,13 +102,9 @@ module Prelude (
     map, (++), filter,
     head, last, tail, init, null, length, (!!),
     reverse,
-    -- ** Reducing lists (folds)
-    foldl, foldl1, foldr, foldr1,
     -- *** Special folds
     and, or, any, all,
-    sum, product,
     concat, concatMap,
-    maximum, minimum,
     -- ** Building lists
     -- *** Scans
     scanl, scanl1, scanr, scanr1,
@@ -96,7 +113,7 @@ module Prelude (
     -- ** Sublists
     take, drop, splitAt, takeWhile, dropWhile, span, break,
     -- ** Searching lists
-    elem, notElem, lookup,
+    notElem, lookup,
     -- ** Zipping and unzipping lists
     zip, zip3, zipWith, zipWith3, unzip, unzip3,
     -- ** Functions on strings
@@ -140,37 +157,15 @@ import System.IO
 import System.IO.Error
 import Data.List
 import Data.Either
+import Data.Foldable    ( Foldable(..) )
 import Data.Maybe
+import Data.Traversable ( Traversable(..) )
 import Data.Tuple
 
-import GHC.Base
+import GHC.Base hiding ( foldr, mapM, sequence )
 import Text.Read
 import GHC.Enum
 import GHC.Num
 import GHC.Real
 import GHC.Float
 import GHC.Show
-
-infixr 0 $!
-
--- -----------------------------------------------------------------------------
--- Miscellaneous functions
-
--- | Strict (call-by-value) application, defined in terms of 'seq'.
-($!)    :: (a -> b) -> a -> b
-f $! x  = let !vx = x in f vx  -- see #2273
-
-#ifdef __HADDOCK__
--- | The value of @'seq' a b@ is bottom if @a@ is bottom, and otherwise
--- equal to @b@.  'seq' is usually introduced to improve performance by
--- avoiding unneeded laziness.
---
--- A note on evaluation order: the expression @seq a b@ does /not/ guarantee
--- that @a@ will be evaluated before @b@. The only guarantee given by @seq@ is
--- that the both @a@ and @b@ will be evaluated before @seq@ returns a value. In
--- particular, this means that @b@ may be evaluated before @a@. If you need to
--- guarantee a specific order of evaluation, you must use the function @pseq@
--- from the parallel package.
-seq :: a -> b -> b
-seq _ y = y
-#endif