+++ rts/RtsFlags.c
@@ -2201,11 +2201,3 @@ void freeRtsArgs(void)
     freeProgArgv();
     freeRtsArgv();
 }
-
-// Local Variables:
-// mode: C
-// fill-column: 80
-// indent-tabs-mode: nil
-// c-basic-offset: 4
-// buffer-file-coding-system: utf-8-unix
-// End: