+++ rts/RtsMessages.c
@@ -470,11 +470,3 @@ parDebugMsgFn(const char *s, va_list ap)
   }
 }
 #endif
-
-// Local Variables:
-// mode: C
-// fill-column: 80
-// indent-tabs-mode: nil
-// c-basic-offset: 4
-// buffer-file-coding-system: utf-8-unix
-// End: