+++ rts/Schedule.c
@@ -3153,11 +3153,3 @@ resurrectThreads (StgTSO *threads)
 	}
     }
 }
-
-// Local Variables:
-// mode: C
-// fill-column: 80
-// indent-tabs-mode: nil
-// c-basic-offset: 4
-// buffer-file-coding-system: utf-8-unix
-// End: