+++ rts/eventlog/EventLog.c
@@ -1220,7 +1220,7 @@ void postThreadLabel(Capability    *cap,
 {
     EventsBuf *eb;
     int strsize = strlen(label);
-    int size = strsize + sizeof(EventCapsetID);
+    int size = strsize + sizeof(EventThreadID);
 
     eb = &capEventBuf[cap->no];
 
@@ -1503,11 +1503,3 @@ void postEventType(EventsBuf *eb, EventType *et)
 }
 
 #endif /* TRACING */
-
-// Local Variables:
-// mode: C
-// fill-column: 80
-// indent-tabs-mode: nil
-// c-basic-offset: 4
-// buffer-file-coding-system: utf-8-unix
-// End: